fix: IPGView reloading on restart, added IPGCog, logging refactors
