+++ alembic/versions/20240226143908_4bcdd5a2bbd5_add_message_id_to_inprogressgame.py
@@ -1,0 +1,33 @@
+"""add message_id to InProgressGame
+
+Revision ID: 4bcdd5a2bbd5
+Revises: b4594ec71cc2
+Create Date: 2024-02-26 14:39:08.079971
+
+"""
+
+import sqlalchemy as sa
+
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision = "4bcdd5a2bbd5"
+down_revision = "b4594ec71cc2"
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    with op.batch_alter_table("in_progress_game", schema=None) as batch_op:
+        batch_op.add_column(sa.Column("message_id", sa.BigInteger(), nullable=True))
+
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    with op.batch_alter_table("in_progress_game", schema=None) as batch_op:
+        batch_op.drop_column("message_id")
+
+    # ### end Alembic commands ###
