+++ alembic/versions/20240227112238_57702a59e6e5_.py
@@ -1,0 +1,25 @@
+"""Merge f283e65712c8 and 4bcdd5a2bbd5
+
+Revision ID: 57702a59e6e5
+Revises: f283e65712c8, 4bcdd5a2bbd5
+Create Date: 2024-02-27 11:22:38.808061
+
+"""
+
+import sqlalchemy as sa
+
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision = "57702a59e6e5"
+down_revision = ("f283e65712c8", "4bcdd5a2bbd5")
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    pass
+
+
+def downgrade():
+    pass
