+++ discord_bots/utils.py
@@ -33,22 +33,18 @@ import discord_bots.config as config
 from discord_bots.bot import bot
 from discord_bots.models import (
     Category,
-    FinishedGame,
-    FinishedGamePlayer,
-    InProgressGame,
-    InProgressGameChannel,
-    InProgressGamePlayer,
     Map,
     MapVote,
     Player,
     PlayerCategoryTrueskill,
     Queue,
-    QueueWaitlist,
     Rotation,
     RotationMap,
     Session,
     SkipMapVote,
 )
+
+_log = logging.getLogger(__name__)
 
 
 MU_LOWER_UNICODE = "\u03BC"
@@ -477,393 +473,8 @@ async def sync(
         try:
             await ctx.bot.tree.sync(guild=guild)
         except discord.HTTPException as e:
-            logging.warn(f"Caught exception trying to sync for guild: ${guild}, e: {e}")
+            _log.warn(f"Caught exception trying to sync for guild: ${guild}, e: {e}")
         else:
             ret += 1
 
     await ctx.send(f"Synced the tree to {ret}/{len(guilds)}.")
-
-
-async def cancel_in_progress_game(interaction: Interaction, game_id: str):
-    session: SQLAlchemySession = Session()
-    game = (
-        session.query(InProgressGame)
-        .filter(InProgressGame.id.startswith(game_id))
-        .first()
-    )
-    if not game:
-        await interaction.response.send_message(
-            embed=Embed(
-                description=f"Could not find game: {game_id}",
-                colour=Colour.red(),
-            )
-        )
-        return
-
-    session.query(InProgressGamePlayer).filter(
-        InProgressGamePlayer.in_progress_game_id == game.id
-    ).delete()
-    for channel in session.query(InProgressGameChannel).filter(
-        InProgressGameChannel.in_progress_game_id == game.id
-    ):
-        if interaction.guild:
-            guild_channel = interaction.guild.get_channel(channel.channel_id)
-            if guild_channel:
-                await guild_channel.delete()
-        session.delete(channel)
-    session.commit()
-
-    session.query(InProgressGame).filter(InProgressGame.id == game.id).delete()
-    session.commit()
-    session.close()
-    await interaction.response.send_message(
-        embed=Embed(
-            description=f"Game {game_id} cancelled",
-            colour=Colour.blue(),
-        )
-    )
-    return
-
-
-async def finish_in_progress_game(
-    interaction: Interaction,
-    outcome: Literal["win", "loss", "tie"],
-    game_id: Optional[str] = None,
-) -> bool:
-    session: sqlalchemy.orm.session.Session = Session()
-    game_player: InProgressGamePlayer | None = (
-        session.query(InProgressGamePlayer)
-        .filter(InProgressGamePlayer.player_id == interaction.user.id)
-        .first()
-    )
-    if not game_player:
-        await interaction.followup.send(
-            embed=discord.Embed(
-                description="You are not in a game!",
-                color=discord.Colour.red(),
-            ),
-            ephemeral=True,
-        )
-        session.close()
-        return False
-
-    in_progress_game: InProgressGame | None
-    if game_id:
-        in_progress_game = (
-            session.query(InProgressGame)
-            .filter(InProgressGame.id == game_player.in_progress_game_id)
-            .filter(InProgressGame.id == game_id)
-            .first()
-        )
-        if not in_progress_game:
-            await interaction.followup.send(
-                embed=discord.Embed(
-                    description="You are not in this game!", color=discord.Colour.red()
-                ),
-                ephemeral=True,
-            )
-            session.close()
-            return False
-    else:
-        in_progress_game = (
-            session.query(InProgressGame)
-            .filter(InProgressGame.id == game_player.in_progress_game_id)
-            .first()
-        )
-        if not in_progress_game:
-            logging.warn(
-                f"No in_progress_game found with id={game_player.in_progress_game_id} for game_player with id={game_player.id}"
-            )
-            await interaction.followup.send(
-                embed=discord.Embed(
-                    description="You are not in a game!", color=discord.Colour.red()
-                ),
-                ephemeral=True,
-            )
-            session.close()
-            return False
-
-    queue: Queue | None = (
-        session.query(Queue).filter(Queue.id == in_progress_game.queue_id).first()
-    )
-    if not queue:
-        # should never happen
-        logging.error(
-            f"Could not find queue with id {in_progress_game.queue_id} for in_progress_game with id {in_progress_game.id}"
-        )
-        await interaction.followup.send(
-            embed=discord.Embed(
-                description="Something went wrong, please contact the server owner",
-                color=discord.Colour.red(),
-            ),
-            ephemeral=True,
-        )
-        session.close()
-        return False
-
-    winning_team = -1
-    if outcome == "win":
-        winning_team = game_player.team
-    elif outcome == "loss":
-        winning_team = (game_player.team + 1) % 2
-    else:
-        # tie
-        winning_team = -1
-
-    players = (
-        session.query(Player)
-        .join(InProgressGamePlayer)
-        .filter(
-            InProgressGamePlayer.player_id == Player.id,
-            InProgressGamePlayer.in_progress_game_id == in_progress_game.id,
-        )
-    ).all()
-    player_ids: list[str] = [player.id for player in players]
-    players_by_id: dict[int, Player] = {player.id: player for player in players}
-    player_category_trueskills_by_id: dict[int, PlayerCategoryTrueskill] = {}
-    if queue.category_id:
-        player_category_trueskills: list[PlayerCategoryTrueskill] = (
-            session.query(PlayerCategoryTrueskill)
-            .filter(
-                PlayerCategoryTrueskill.player_id.in_(player_ids),
-                PlayerCategoryTrueskill.category_id == queue.category_id,
-            )
-            .all()
-        )
-        player_category_trueskills_by_id = {
-            pct.player_id: pct for pct in player_category_trueskills
-        }
-    in_progress_game_players = (
-        session.query(InProgressGamePlayer)
-        .filter(InProgressGamePlayer.in_progress_game_id == in_progress_game.id)
-        .all()
-    )
-    team0_rated_ratings_before = []
-    team1_rated_ratings_before = []
-    team0_players: list[InProgressGamePlayer] = []
-    team1_players: list[InProgressGamePlayer] = []
-    for in_progress_game_player in in_progress_game_players:
-        player = players_by_id[in_progress_game_player.player_id]
-        if in_progress_game_player.team == 0:
-            team0_players.append(in_progress_game_player)
-            if player.id in player_category_trueskills_by_id:
-                pct = player_category_trueskills_by_id[player.id]
-                team0_rated_ratings_before.append(Rating(pct.mu, pct.sigma))
-            else:
-                team0_rated_ratings_before.append(
-                    Rating(player.rated_trueskill_mu, player.rated_trueskill_sigma)
-                )
-        else:
-            team1_players.append(in_progress_game_player)
-            if player.id in player_category_trueskills_by_id:
-                pct = player_category_trueskills_by_id[player.id]
-                team1_rated_ratings_before.append(Rating(pct.mu, pct.sigma))
-            else:
-                team1_rated_ratings_before.append(
-                    Rating(player.rated_trueskill_mu, player.rated_trueskill_sigma)
-                )
-
-    if queue.category_id:
-        category: Category | None = (
-            session.query(Category).filter(Category.id == queue.category_id).first()
-        )
-        if category:
-            category_name = category.name
-        else:
-            # should never happen
-            logging.error(
-                f"Could not find category with id {queue.category_id} for queue with id {queue.id}"
-            )
-            await interaction.followup.send(
-                embed=discord.Embed(
-                    description="Something went wrong, please contact the server owner",
-                    color=discord.Colour.red(),
-                ),
-                ephemeral=True,
-            )
-            session.close()
-            return False
-    else:
-        category_name = None
-
-    finished_game = FinishedGame(
-        average_trueskill=in_progress_game.average_trueskill,
-        finished_at=datetime.now(timezone.utc),
-        game_id=in_progress_game.id,
-        is_rated=queue.is_rated,
-        map_full_name=in_progress_game.map_full_name,
-        map_short_name=in_progress_game.map_short_name,
-        queue_name=queue.name,
-        category_name=category_name,
-        started_at=in_progress_game.created_at,
-        team0_name=in_progress_game.team0_name,
-        team1_name=in_progress_game.team1_name,
-        win_probability=in_progress_game.win_probability,
-        winning_team=winning_team,
-    )
-    session.add(finished_game)
-
-    result = None
-    if winning_team == -1:
-        result = [0, 0]
-    elif winning_team == 0:
-        result = [0, 1]
-    elif winning_team == 1:
-        result = [1, 0]
-
-    team0_rated_ratings_after: list[Rating]
-    team1_rated_ratings_after: list[Rating]
-    if len(players) > 1:
-        team0_rated_ratings_after, team1_rated_ratings_after = rate(
-            [team0_rated_ratings_before, team1_rated_ratings_before], result
-        )
-    else:
-        # Mostly useful for creating solo queues for testing, no real world
-        # application
-        team0_rated_ratings_after, team1_rated_ratings_after = (
-            team0_rated_ratings_before,
-            team1_rated_ratings_before,
-        )
-
-    for i, team0_gip in enumerate(team0_players):
-        player = players_by_id[team0_gip.player_id]
-        finished_game_player = FinishedGamePlayer(
-            finished_game_id=finished_game.id,
-            player_id=player.id,
-            player_name=player.name,
-            team=team0_gip.team,
-            rated_trueskill_mu_before=team0_rated_ratings_before[i].mu,
-            rated_trueskill_sigma_before=team0_rated_ratings_before[i].sigma,
-            rated_trueskill_mu_after=team0_rated_ratings_after[i].mu,
-            rated_trueskill_sigma_after=team0_rated_ratings_after[i].sigma,
-        )
-        trueskill_rating = team0_rated_ratings_after[i]
-        # Regardless of category, always update the master trueskill. That way
-        # when we create new categories off of it the data isn't completely
-        # stale
-        player.rated_trueskill_mu = trueskill_rating.mu
-        player.rated_trueskill_sigma = trueskill_rating.sigma
-        if player.id in player_category_trueskills_by_id:
-            pct = player_category_trueskills_by_id[player.id]
-            pct.mu = trueskill_rating.mu
-            pct.sigma = trueskill_rating.sigma
-            pct.rank = trueskill_rating.mu - 3 * trueskill_rating.sigma
-        else:
-            session.add(
-                PlayerCategoryTrueskill(
-                    player_id=player.id,
-                    category_id=queue.category_id,
-                    mu=trueskill_rating.mu,
-                    sigma=trueskill_rating.sigma,
-                    rank=trueskill_rating.mu - 3 * trueskill_rating.sigma,
-                )
-            )
-        session.add(finished_game_player)
-    for i, team1_gip in enumerate(team1_players):
-        player = players_by_id[team1_gip.player_id]
-        finished_game_player = FinishedGamePlayer(
-            finished_game_id=finished_game.id,
-            player_id=player.id,
-            player_name=player.name,
-            team=team1_gip.team,
-            rated_trueskill_mu_before=team1_rated_ratings_before[i].mu,
-            rated_trueskill_sigma_before=team1_rated_ratings_before[i].sigma,
-            rated_trueskill_mu_after=team1_rated_ratings_after[i].mu,
-            rated_trueskill_sigma_after=team1_rated_ratings_after[i].sigma,
-        )
-        trueskill_rating = team1_rated_ratings_after[i]
-        # Regardless of category, always update the master trueskill. That way
-        # when we create new categories off of it the data isn't completely
-        # stale
-        player.rated_trueskill_mu = trueskill_rating.mu
-        player.rated_trueskill_sigma = trueskill_rating.sigma
-        if player.id in player_category_trueskills_by_id:
-            pct = player_category_trueskills_by_id[player.id]
-            pct.mu = trueskill_rating.mu
-            pct.sigma = trueskill_rating.sigma
-            pct.rank = trueskill_rating.mu - 3 * trueskill_rating.sigma
-        else:
-            session.add(
-                PlayerCategoryTrueskill(
-                    player_id=player.id,
-                    category_id=queue.category_id,
-                    mu=trueskill_rating.mu,
-                    sigma=trueskill_rating.sigma,
-                    rank=trueskill_rating.mu - 3 * trueskill_rating.sigma,
-                )
-            )
-        session.add(finished_game_player)
-
-    session.query(InProgressGamePlayer).filter(
-        InProgressGamePlayer.in_progress_game_id == in_progress_game.id
-    ).delete()
-    in_progress_game.is_finished = True
-    embed_description = ""
-    duration: timedelta = finished_game.finished_at.replace(
-        tzinfo=timezone.utc
-    ) - in_progress_game.created_at.replace(tzinfo=timezone.utc)
-    if winning_team == 0:
-        embed_description = f"**Winner:** {in_progress_game.team0_name}\n**Duration:** {duration.seconds // 60} minutes"
-    elif winning_team == 1:
-        embed_description = f"**Winner:** {in_progress_game.team1_name}\n**Duration:** {duration.seconds // 60} minutes"
-    else:
-        embed_description = (
-            f"**Tie game**\n**Duration:** {duration.seconds // 60} minutes"
-        )
-    session.add(
-        QueueWaitlist(
-            channel_id=interaction.channel_id,  # not sure about this column and what it's used for
-            finished_game_id=finished_game.id,
-            guild_id=interaction.guild_id,
-            in_progress_game_id=in_progress_game.id,
-            queue_id=queue.id,
-            end_waitlist_at=datetime.now(timezone.utc)
-            + timedelta(seconds=config.RE_ADD_DELAY),
-        )
-    )
-
-    # Reward raffle tickets
-    reward = (
-        session.query(RotationMap.raffle_ticket_reward)
-        .join(Map, Map.id == RotationMap.map_id)
-        .join(Rotation, Rotation.id == RotationMap.rotation_id)
-        .join(Queue, Queue.rotation_id == Rotation.id)
-        .filter(Map.short_name == in_progress_game.map_short_name)
-        .filter(Queue.id == in_progress_game.queue_id)
-        .scalar()
-    )
-    if reward == 0:
-        reward = config.DEFAULT_RAFFLE_VALUE
-
-    for player in players:
-        player.raffle_tickets += reward
-        session.add(player)
-
-    queue_name = queue.name
-    short_in_progress_game_id = in_progress_game.id.split("-")[0]
-    await interaction.followup.send(
-        embed=discord.Embed(
-            title=f"Game '{queue_name}' ({short_in_progress_game_id}) finished",
-            description=embed_description,
-            colour=discord.Colour.green(),
-        )
-    )
-    if config.CHANNEL_ID and config.CHANNEL_ID != interaction.channel_id:
-        channel: discord.channel.TextChannel | None = discord.utils.get(
-            interaction.guild.text_channels, id=config.CHANNEL_ID
-        )
-        if channel:
-            await channel.send(
-                embed=discord.Embed(
-                    title=f"Game '{queue_name}' ({short_in_progress_game_id}) finished",
-                    description=embed_description,
-                    colour=discord.Colour.green(),
-                )
-            )
-            await upload_stats_screenshot_imgkit_channel(channel)
-    else:
-        await upload_stats_screenshot_imgkit_interaction(interaction)
-
-    session.commit()
-    session.close()
-    return True
